Updating participant view SQL to include new sample and order sample statuses.
