--- rest-api/alembic/versions/f81a8b941a77_fix_pv_column_name.py
@@ -1,185 +1,0 @@
-"""fix pv column name
-
-Revision ID: f81a8b941a77
-Revises: 29e344f31b21
-Create Date: 2018-01-17 09:40:19.259683
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import model.utils
-
-
-from participant_enums import PhysicalMeasurementsStatus, QuestionnaireStatus, OrderStatus
-from participant_enums import WithdrawalStatus, SuspensionStatus
-from participant_enums import EnrollmentStatus, Race, SampleStatus, OrganizationType
-from participant_enums import MetricSetType, MetricsKey
-from model.site_enums import SiteStatus
-from model.code import CodeType
-
-# revision identifiers, used by Alembic.
-revision = 'f81a8b941a77'
-down_revision = '29e344f31b21'
-branch_labels = None
-depends_on = None
-
-
-_PARTICIPANT_VIEW_SQL = """
-CREATE OR REPLACE VIEW participant_view AS
- SELECT
-   p.participant_id,
-   p.sign_up_time,
-   p.withdrawal_status,
-   p.withdrawal_time,
-   p.suspension_status,
-   p.suspension_time,
-   hpo.name hpo,
-   ps.zip_code,
-   state_code.value state,
-   recontact_method_code.value recontact_method,
-   language_code.value language,
-   TIMESTAMPDIFF(YEAR, ps.date_of_birth, CURDATE()) age_years,
-   gender_code.value gender,
-   sex_code.value sex,
-   sexual_orientation_code.value sexual_orientation,
-   education_code.value education,
-   income_code.value income,
-   ps.enrollment_status,
-   ps.race,
-   ps.physical_measurements_status,
-   ps.physical_measurements_finalized_time,
-   ps.physical_measurements_time,
-   ps.physical_measurements_created_site_id,
-   ps.physical_measurements_finalized_site_id,
-   ps.consent_for_study_enrollment,
-   ps.consent_for_study_enrollment_time,
-   ps.consent_for_electronic_health_records,
-   ps.consent_for_electronic_health_records_time,
-   ps.questionnaire_on_overall_health,
-   ps.questionnaire_on_overall_health_time,
-   ps.questionnaire_on_lifestyle,
-   ps.questionnaire_on_lifestyle_time,
-   ps.questionnaire_on_the_basics,
-   ps.questionnaire_on_the_basics_time,
-   ps.questionnaire_on_healthcare_access,
-   ps.questionnaire_on_healthcare_access_time,
-   ps.questionnaire_on_medical_history,
-   ps.questionnaire_on_medical_history_time,
-   ps.questionnaire_on_medications,
-   ps.questionnaire_on_medications_time,
-   ps.questionnaire_on_family_health,
-   ps.questionnaire_on_family_health_time,
-   ps.biospecimen_status,
-   ps.biospecimen_order_time,
-   ps.biospecimen_source_site_id,
-   ps.biospecimen_collected_site_id,
-   ps.biospecimen_processed_site_id,
-   ps.biospecimen_finalized_site_id,
-   ps.sample_order_status_1sst8,
-   ps.sample_order_status_1sst8_time,
-   ps.sample_order_status_1pst8,
-   ps.sample_order_status_1pst8_time,
-   ps.sample_order_status_1hep4,
-   ps.sample_order_status_1hep4_time,
-   ps.sample_order_status_1ed04,
-   ps.sample_order_status_1ed04_time,
-   ps.sample_order_status_1ed10,
-   ps.sample_order_status_1ed10_time,
-   ps.sample_order_status_2ed10,
-   ps.sample_order_status_2ed10_time,
-   ps.sample_order_status_1ur10,
-   ps.sample_order_status_1ur10_time,
-   ps.sample_order_status_1sal,
-   ps.sample_order_status_1sal_time,
-   ps.sample_status_1sst8,
-   ps.sample_status_1sst8_time,
-   ps.sample_status_1pst8,
-   ps.sample_status_1pst8_time,
-   ps.sample_status_1hep4,
-   ps.sample_status_1hep4_time,
-   ps.sample_status_1ed04,
-   ps.sample_status_1ed04_time,
-   ps.sample_status_1ed10,
-   ps.sample_status_1ed10_time,
-   ps.sample_status_2ed10,
-   ps.sample_status_2ed10_time,
-   ps.sample_status_1ur10,
-   ps.sample_status_1ur10_time,
-   ps.sample_status_1sal,
-   ps.sample_status_1sal_time,
-   ps.num_completed_baseline_ppi_modules,
-   ps.num_completed_ppi_modules,
-   ps.num_baseline_samples_arrived,
-   ps.samples_to_isolate_dna,
-   ps.consent_for_cabor,
-   ps.consent_for_cabor_time,
-   (SELECT IFNULL(GROUP_CONCAT(
-      IF(ac.value = 'WhatRaceEthnicity_RaceEthnicityNoneOfThese',
-         'NoneOfThese',
-         TRIM(LEADING 'WhatRaceEthnicity_' FROM
-              TRIM(LEADING 'PMI_' FROM ac.value)))),
-    'None')
-    FROM questionnaire_response qr, questionnaire_response_answer qra,
-      questionnaire_question qq, code c, code ac
-    WHERE qra.end_time IS NULL AND
-          qr.questionnaire_response_id = qra.questionnaire_response_id AND
-          qra.question_id = qq.questionnaire_question_id AND
-          qq.code_id = c.code_id AND c.value = 'Race_WhatRaceEthnicity' AND
-          qr.participant_id = p.participant_id AND
-          qra.value_code_id = ac.code_id AND
-          ac.value != 'WhatRaceEthnicity_Hispanic'
-   ) race_codes,
-   (SELECT COUNT(ac.value)
-    FROM questionnaire_response qr, questionnaire_response_answer qra,
-      questionnaire_question qq, code c, code ac
-    WHERE qra.end_time IS NULL AND
-      qr.questionnaire_response_id = qra.questionnaire_response_id AND
-      qra.question_id = qq.questionnaire_question_id AND
-      qq.code_id = c.code_id AND c.value = 'Race_WhatRaceEthnicity' AND
-      qr.participant_id = p.participant_id AND
-      qra.value_code_id = ac.code_id AND
-      ac.value = 'WhatRaceEthnicity_Hispanic'
-   ) hispanic
- FROM
-   participant p
-     LEFT OUTER JOIN hpo ON p.hpo_id = hpo.hpo_id
-     LEFT OUTER JOIN participant_summary ps ON p.participant_id = ps.participant_id
-     LEFT OUTER JOIN code state_code ON ps.state_id = state_code.code_id
-     LEFT OUTER JOIN code recontact_method_code ON ps.recontact_method_id = recontact_method_code.code_id
-     LEFT OUTER JOIN code language_code ON ps.language_id = language_code.code_id
-     LEFT OUTER JOIN code gender_code ON ps.gender_identity_id = gender_code.code_id
-     LEFT OUTER JOIN code sex_code ON ps.sex_id = sex_code.code_id
-     LEFT OUTER JOIN code sexual_orientation_code ON ps.sexual_orientation_id = sexual_orientation_code.code_id
-     LEFT OUTER JOIN code education_code ON ps.education_id = education_code.code_id
-     LEFT OUTER JOIN code income_code ON ps.income_id = income_code.code_id
-     WHERE (ps.email IS NULL OR ps.email NOT LIKE '%@example.com') AND
-           (hpo.name IS NULL OR hpo.name != 'TEST')
-"""
-
-def upgrade(engine_name):
-    globals()["upgrade_%s" % engine_name]()
-
-
-def downgrade(engine_name):
-    globals()["downgrade_%s" % engine_name]()
-
-
-
-def upgrade_rdr():
-    op.execute(_PARTICIPANT_VIEW_SQL)
-
-
-def downgrade_rdr():
-    pass
-
-
-def upgrade_metrics():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade_metrics():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
